requirements.txt
asgiref
Django
numpy
Pillow
pytesseract
sqlparse
tzdata
fpdf
mysqlclient